青福换电

+++ src/main/java/com/xiliulou/electricity/entity/EleDepositOrder.java
@@ -60,7 +60,7 @@ public class EleDepositOrder {
     //租户id
     private Integer tenantId;
 
-    private Integer franchiseeId;
+    private Long franchiseeId;
 
     public static final Integer STATUS_INIT = 0;
     public static final Integer STATUS_SUCCESS = 1;
